refactor, reorganization

+++ stop_motion.py
@@ -1,19 +1,17 @@
 #!/usr/bin/python3
-import pygame
 import glob
 import os
 import sys
 import subprocess
+import argparse
 
-# from pygame.locals import *
 import ffmpeg
 from gpiozero import Button
 from picamera import PiCamera
 from PIL import Image
 
 HOME_DIR = "/home/pi/Documents/stopmo_files"
-PROJECT = sys.argv[1]
-PAD_WIDTH = 2
+PAD_WIDTH = 4
 
 stop_preview_button = Button(2)
 preview_button = Button(3)
@@ -23,19 +21,7 @@ preview_film_button = Button(27)
 exit_program_button = Button(22)
 
 
-# pygame.init()
-# W, H = pygame.display.list_modes()[0]
-# WIDTH = (int(W*.9) // 32) * 32
-# HEIGHT = (int(H*.9) // 32) * 32
-# X_OFFSET = W - WIDTH // 2
-# Y_OFFSET = H - HEIGHT // 2
-# print('Resolution: ({WIDTH}, {HEIGHT}).'.format(WIDTH=WIDTH, HEIGHT=HEIGHT))
-# SCREEN = pygame.display.set_mode([WIDTH, HEIGHT])
-# pygame.mouse.set_visible = False
-# pygame.display.toggle_fullscreen()
 CAMERA = PiCamera(sensor_mode=2)
-# CAMERA.preview_alpha = 128
-# CAMERA.resolution = (WIDTH, HEIGHT)
 
 
 def _pad(resolution, width=32, height=16):
@@ -43,6 +29,8 @@ def _pad(resolution, width=32, height=16):
     # up to the nearest multiple of *width* and *height*; this is
     # needed because overlays require padding to the camera's
     # block size (32x16)
+    if debug_mode:
+        print("_pad function called")
     return (
         ((resolution[0] + (width - 1)) // width) * width,
         ((resolution[1] + (height - 1)) // height) * height,
@@ -50,14 +38,17 @@ def _pad(resolution, width=32, height=16):
 
 
 def remove_overlays(camera):
-
     # Remove all overlays from the camera preview
+    if debug_mode:
+        print("remove_overlays function called")
     for o in camera.overlays:
         camera.remove_overlay(o)
 
 
 def preview_overlay(camera=None):
     # Remove all overlays
+    if debug_mode:
+        print("preview_overlay function called")
     remove_overlays(camera)
 
     # Get an Image object of the chosen overlay
@@ -71,22 +62,21 @@ def preview_overlay(camera=None):
     camera.add_overlay(pad.tobytes(), alpha=128, layer=3)
 
 
-def echo(stmt):
-    print(stmt)
-
-
 def stop():
-    print("stop preview button pressed")
+    if debug_mode():
+        print("stop preview button pressed")
+        return
     remove_overlays(CAMERA)
     CAMERA.stop_preview()
 
 
 def preview():
-    print("preview button pressed")
+    if debug_mode:
+        print("preview button pressed")
+        return
     CAMERA.start_preview()
-    frame_base_str = '{}/{}/frames/frame_{}.jpg'
-    if get_next_frame(offset=0) == frame_base_str.format(HOME_DIR,
-                                                         PROJECT,
+    frame_base_str = '{}/frame_{}.jpg'
+    if get_next_frame(offset=0) == frame_base_str.format(frames_dir,
                                                          str(0).zfill(
                                                                       PAD_WIDTH
                                                                       )
@@ -97,45 +87,57 @@ def preview():
 
 
 def deleteframe_button():
-    echo("delete frame button pressed")
+    if debug_mode:
+        print("delete frame button pressed")
+        return
     os.remove(get_next_frame(offset=0))
 
 
 def take_picture():
-    print("take picture button pressed")
+    if debug_mode:
+        print("take picture button pressed")
+        return
     frame_fname = get_next_frame()
     CAMERA.capture(frame_fname)
     ghost_preview()
 
 
 def ghost_preview():
-    print("ghost button pressed")
+    if debug_mode:
+        print("ghost_preview function called")
     preview_overlay(CAMERA)
 
 
 def exit_button():
+    if debug_mode:
+        print("exit button pushed")
     exit()
 
 
 def get_next_frame(offset=1):
-    frames = glob.glob("{}/{}/frames/frame_*.jpg".format(HOME_DIR, PROJECT))
+    if debug_mode:
+        print("get_next_frame function called")
+    frames = glob.glob("{}/frame_*.jpg".format(frames_dir))
     if len(frames) == 0:
-        return '{}/{}/frames/frame_{}.jpg'.format(HOME_DIR, PROJECT,
-                                                  str(1).zfill(PAD_WIDTH))
+        return '{}/frame_{}.jpg'.format(frames_dir,
+                                        str(1).zfill(PAD_WIDTH))
     sequence = [int(os.path.basename(x).split(".")[0].split("_")[-1])
                 for x
                 in frames]
-    return '{}/{}/frames/frame_{}.jpg'.format(HOME_DIR, PROJECT,
-                                              str(
-                                                  max(sequence)+offset
-                                                  ).zfill(PAD_WIDTH))
+    return '{}/frame_{}.jpg'.format(frames_dir, PROJECT,
+                                    str(
+                                        max(sequence)+offset
+                                        ).zfill(PAD_WIDTH))
 
 
 def assemble_and_preview():
+    if debug_mode:
+        print("assemble_and_preview button pushed")
+        return
     stop()
-    output_fname = '{hdir}/{proj}/movie/{proj}.mp4'.format(hdir=HOME_DIR,
-                                                            proj=PROJECT)
-    video_in = ffmpeg.input('{}/{}/frame_*.jpg'.format(HOME_DIR, PROJECT),
+    output_fname = '{m_dir}/{proj}.mp4'.format(m_dir=movie_dir,
+                                               proj=PROJECT)
+    video_in = ffmpeg.input('{}/frame_*.jpg'.format(frames_dir),
                             pattern_type='glob',
                             framerate=12)
     video_out = ffmpeg.output(video_in, output_fname)
@@ -154,7 +156,27 @@ def main():
 
 
 if __name__ == '__main__':
-    if not os.path.exists("{}/{}/frames".format(HOME_DIR, PROJECT)):
-        os.makedirs("{}/{}/frames".format(HOME_DIR, PROJECT))
-        os.mkdir("{}/{}/movie".format(HOME_DIR, PROJECT))
+    parser = argparse.ArgumentParser(description="Stop Motion Pi")
+    parser.add_argument('project_name',
+                        type=str,
+                        help="name for project, no spaces, used for directory")
+    parser.add_argument('--debug_mode', '-d',
+                        action='store_true',
+                        help="debug functions by just printing" +
+                             "what is to be done to standard output")
+    args = parser.parse_args()
+    debug_mode = args['debug_mode']
+    PROJECT = args['project_name']
+
+    frames_dir = "{}/{}/frames".format(HOME_DIR, PROJECT)
+    movie_dir = "{}/{}/movie".format(HOME_DIR, PROJECT)
+
+    if not os.path.exists(frames_dir):
+        if debug_mode:
+            print("creating 'frames' directory at {}".format(frames_dir))
+        os.makedirs(frames_dir)
+    if not os.path.exists(movie_dir):
+        if debug_mode:
+            print("creating 'movie' directory at {}".format(movie_dir))
+        os.makedirs(movie_dir)
     main()
